перенес полезный метод в Trees
